Flytter test for å spare på byggetid.
